Refactor Spring configuration #3
