--- spring-boot-integration/pg-index-health-test-starter/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=io.github.mfvanek.pg.spring.DatabaseStructureHealthAutoConfiguration
