+++ spring-boot-integration/pg-index-health-test-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,2 @@
 io.github.mfvanek.pg.spring.DatabaseStructureHealthAutoConfiguration
+io.github.mfvanek.pg.spring.DatabaseStructureChecksAutoConfiguration
