+++ spring-boot-integration/pg-index-health-test-starter/src/test/java/io/github/mfvanek/pg/spring/AutoConfigurationTestBase.java
@@ -68,7 +68,7 @@ abstract class AutoConfigurationTestBase {
 
     @Nonnull
     protected ApplicationContextRunner assertWithTestConfig() {
-        return contextRunner.withUserConfiguration(DatabaseStructureHealthAutoConfiguration.class);
+        return contextRunner.withUserConfiguration(DatabaseStructureHealthAutoConfiguration.class, DatabaseStructureChecksAutoConfiguration.class);
     }
 
     protected static <C extends ConfigurableApplicationContext> void initialize(@Nonnull final C applicationContext) {
